refactor for Component filename
