+++ src/components/PTO-Table/PTO-Table.js
@@ -8,9 +8,9 @@ import * as PTOConstants from '../../constants/pto-constants';
 import Table from '../A10UI/Table/Table';
 import Th from '../A10UI/Table/Th';
 import Td from '../A10UI/Table/Td';
-import DeleteButton from '../A10UI/Button/DeleteButton';
-import ApproveButton from '../A10UI/Button/ApproveButton';
-import DenyButton from '../A10UI/Button/DenyButton';
+import DeleteButton from '../A10UI/Button/Delete-Button';
+import ApproveButton from '../A10UI/Button/Approve-Button';
+import DenyButton from '../A10UI/Button/Deny-Button';
 
 let TableHeaders = ({ titleKeyMap, onSortHandler, sortBy, enableSort }) => {
     let headerHtml = titleKeyMap.map((headerObj, index) => {
